backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pymupdf
python-pptx
langchain
langchain-openai
langchain-google-genai
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pandas
matplotlib
scipy
lxml
beautifulsoup4
pytest
python-docx
google-generativeai
pip-tools
